+++ src/route/reports.js
@@ -1,12 +1,12 @@
 const express = require("express");
 const { body } = require("express-validator");
-const validator = require("../middleware/validator");
+const validator = require("../middlewares/validator");
 
 const router = express.Router();
 const reportHandlers = require("../service/reports");
 
 // 라우터 접근 시 로그인 필요
-router.use(require("../middleware/auth"));
+router.use(require("../middlewares/auth"));
 
 router.post(
   "/create",
