+++ src/route/rooms.js
@@ -3,11 +3,11 @@ const { query, body } = require("express-validator");
 const router = express.Router();
 
 const roomHandlers = require("../service/rooms");
-const validator = require("../middleware/validator");
+const validator = require("../middlewares/validator");
 const patterns = require("../db/patterns");
 
 // 라우터 접근 시 로그인 필요
-router.use(require("../middleware/auth"));
+router.use(require("../middlewares/auth"));
 
 // 특정 id 방 세부사항 보기
 router.get(
